Use int for audio sliders, show current value (#235)

+++ include/dusk/settings.h
@@ -15,11 +15,11 @@ struct UserSettings {
 
     struct {
         // Audio
-        float masterVolume;
-        float mainMusicVolume;
-        float subMusicVolume;
-        float soundEffectsVolume;
-        float fanfareVolume;
+        int masterVolume;
+        int mainMusicVolume;
+        int subMusicVolume;
+        int soundEffectsVolume;
+        int fanfareVolume;
         bool enableReverb;
     } audio;
